+++ ali/common/src/main/java/com/yanny/ali/plugin/Plugin.java
@@ -136,6 +136,8 @@ public class Plugin implements IPlugin {
     public void registerServer(IServerRegistry registry) {
         new CommonValueTooltip<IServerUtils, IServerRegistry>().registerAll(registry);
 
+        EnumTypes.TRANSLATED_ENUMS.forEach(registry::registerEnumTranslation);
+
         registry.registerItemCollector(LootItem.class, ItemCollectorUtils::collectItems);
         registry.registerItemCollector(TagEntry.class, ItemCollectorUtils::collectTags);
         registry.registerItemCollector(AlternativesEntry.class, ItemCollectorUtils::collectComposite);
@@ -401,6 +403,7 @@ public class Plugin implements IPlugin {
         registry.registerValueTooltip(DataComponentPredicate.Type.class, RegistriesTooltipUtils::getDataComponentPredicateTypeTooltip);
         registry.registerValueTooltip(ZombieNautilusVariant.class, RegistriesTooltipUtils::getZombieNautilusVariantTooltip);
 
+        registry.registerValueTooltip(Enum.class, ValueTooltipUtils::getEnumTooltip);
         registry.registerValueTooltip(LootItemCondition.class, ValueTooltipUtils::getConditionTooltip);
         registry.registerValueTooltip(LootItemFunction.class, ValueTooltipUtils::getFunctionTooltip);
         registry.registerValueTooltip(Ingredient.class, ValueTooltipUtils::getIngredientTooltip);