+++ ali/common/src/main/java/com/yanny/ali/plugin/server/ValueTooltipUtils.java
@@ -50,6 +50,11 @@ import static com.yanny.ali.plugin.server.GenericTooltipUtils.getTranslationKey;
 
 public class ValueTooltipUtils {
     @NotNull
+    public static TooltipBuilder getEnumTooltip(IServerUtils utils, Enum<?> condition) {
+        return utils.getEnumTranslation(utils, condition);
+    }
+
+    @NotNull
     public static TooltipBuilder getConditionTooltip(IServerUtils utils, LootItemCondition condition) {
         return utils.getConditionTooltip(utils, condition);
     }