+++ ali/common/src/test/java/com/yanny/ali/test/DataComponentTooltipTest.java
@@ -114,7 +114,7 @@ public class DataComponentTooltipTest {
 
     @Test
     public void testRarityTooltip() {
-        assertTooltip(DataComponentTooltipUtils.getRarityTooltip(UTILS, Rarity.EPIC).build(), List.of("Rarity: EPIC"));
+        assertTooltip(DataComponentTooltipUtils.getRarityTooltip(UTILS, Rarity.EPIC).build(), List.of("Rarity: Epic"));
     }
 
     @Test
@@ -186,15 +186,15 @@ public class DataComponentTooltipTest {
                 "    -> Modifier:",
                 "      -> Id: minecraft:hello",
                 "      -> Amount: 0.5",
-                "      -> Operation: ADD_VALUE",
-                "    -> Slot: HEAD",
+                "      -> Operation: Add Value",
+                "    -> Slot: Head",
                 "  -> Entry:",
                 "    -> Attribute: minecraft:armor_toughness",
                 "    -> Modifier:",
                 "      -> Id: minecraft:world",
                 "      -> Amount: 1.25",
-                "      -> Operation: ADD_MULTIPLIED_TOTAL",
-                "    -> Slot: HAND"
+                "      -> Operation: Multiply Total",
+                "    -> Slot: Hand"
         ));
     }
 
@@ -484,7 +484,7 @@ public class DataComponentTooltipTest {
 
     @Test
     public void testMapPostProcessingTooltip() {
-        assertTooltip(DataComponentTooltipUtils.getMapPostProcessingTooltip(UTILS, MapPostProcessing.SCALE).build(), List.of("Value: SCALE"));
+        assertTooltip(DataComponentTooltipUtils.getMapPostProcessingTooltip(UTILS, MapPostProcessing.SCALE).build(), List.of("Value: Scale"));
     }
 
     @Test
@@ -731,7 +731,7 @@ public class DataComponentTooltipTest {
                 true,
                 false
         )).build(), List.of(
-                "Shape: LARGE_BALL",
+                "Shape: Large Ball",
                 "Colors: [1, 2, 3]",
                 "Fade Colors: []",
                 "Has Trail: true",
@@ -751,13 +751,13 @@ public class DataComponentTooltipTest {
                 "Flight Duration: 10",
                 "Explosions:",
                 "  -> Entry:",
-                "    -> Shape: STAR",
+                "    -> Shape: Star",
                 "    -> Colors: []",
                 "    -> Fade Colors: []",
                 "    -> Has Trail: true",
                 "    -> Has Twinkle: true",
                 "  -> Entry:",
-                "    -> Shape: CREEPER",
+                "    -> Shape: Creeper",
                 "    -> Colors: []",
                 "    -> Fade Colors: []",
                 "    -> Has Trail: true",
@@ -778,15 +778,15 @@ public class DataComponentTooltipTest {
         ))).build(), List.of(
                 "Banner Patterns:",
                 "  -> minecraft:base",
-                "    -> Color: BLUE",
+                "    -> Color: Blue",
                 "  -> minecraft:border",
-                "    -> Color: RED"
+                "    -> Color: Red"
         ));
     }
 
     @Test
     public void testDyeColorTooltip() {
-        assertTooltip(DataComponentTooltipUtils.getDyeColorTooltip(UTILS, DyeColor.CYAN).build(), List.of("Color: CYAN"));
+        assertTooltip(DataComponentTooltipUtils.getDyeColorTooltip(UTILS, DyeColor.CYAN).build(), List.of("Color: Cyan"));
     }
 
     @Test