+++ ali/common/src/test/java/com/yanny/ali/test/GenericTooltipTest.java
@@ -167,14 +167,14 @@ public class GenericTooltipTest {
                 .build()).build(Lang.Branch.MODIFIER), List.of(
                 "Modifier:",
                 "  -> Attribute: minecraft:armor",
-                "  -> Operation: ADD_MULTIPLIED_TOTAL",
+                "  -> Operation: Multiply Total",
                 "  -> Amount: 1-5",
                 "  -> Id: minecraft:armor",
                 "  -> Equipment Slots:",
-                "    -> FEET",
-                "    -> LEGS",
-                "    -> CHEST",
-                "    -> HEAD"
+                "    -> Feet",
+                "    -> Legs",
+                "    -> Chest",
+                "    -> Head"
         ));
     }
 
@@ -553,7 +553,7 @@ public class GenericTooltipTest {
                 "  -> Components:",
                 "    -> Expected Components:",
                 "      -> minecraft:base_color",
-                "        -> Color: BLUE"
+                "        -> Color: Blue"
         ));
         assertTooltip(ValueTooltipUtils.getItemPredicateTooltip(UTILS, ItemPredicate.Builder.item()
                 .of(LOOKUP.lookupOrThrow(Registries.ITEM), Items.CAKE, Items.NETHERITE_AXE)
@@ -818,7 +818,7 @@ public class GenericTooltipTest {
     @Test
     public void testGameTypeTooltip() {
         assertTooltip(ValueTooltipUtils.getGameTypePredicateTooltip(UTILS, GameTypePredicate.of(GameType.SPECTATOR)).build(Lang.Branch.GAME_TYPES), List.of(
-                "Game Type: SPECTATOR"
+                "Game Type: Spectator"
         ));
     }
 
@@ -852,11 +852,11 @@ public class GenericTooltipTest {
     @Test
     public void testListOperationTooltip() {
         assertTooltip(ValueTooltipUtils.getListOperationTooltip(UTILS, new ListOperation.Insert(1)).build(Lang.Value.LIST_OPERATION), List.of(
-                "List Operation: INSERT",
+                "List Operation: Insert",
                 "  -> Offset: 1"
         ));
         assertTooltip(ValueTooltipUtils.getListOperationTooltip(UTILS, new ListOperation.ReplaceSection(1, Optional.of(2))).build(Lang.Value.LIST_OPERATION), List.of(
-                "List Operation: REPLACE_SECTION",
+                "List Operation: Replace Section",
                 "  -> Offset: 1",
                 "  -> Size: 2"
         ));
@@ -916,7 +916,7 @@ public class GenericTooltipTest {
                 Optional.of(false)
         )).build(Lang.Branch.CONDITION), List.of(
                 "Predicate:",
-                "  -> Shape: CREEPER",
+                "  -> Shape: Creeper",
                 "  -> Trail: false",
                 "  -> Twinkle: true"
         ));
@@ -943,8 +943,8 @@ public class GenericTooltipTest {
                 "    -> minecraft:gravity",
                 "  -> Id: minecraft:test",
                 "  -> Amount: 1.5-3.1",
-                "  -> Operation: ADD_VALUE",
-                "  -> Slot: ARMOR"
+                "  -> Operation: Add Value",
+                "  -> Slot: Armor"
         ));
     }
 
@@ -959,7 +959,7 @@ public class GenericTooltipTest {
         ).build(Lang.Branch.COMPONENTS), List.of(
                 "Components:",
                 "  -> minecraft:base_color",
-                "    -> Color: BLUE",
+                "    -> Color: Blue",
                 "  -> minecraft:custom_name",
                 "    -> Custom Name: Hello",
                 "  -> minecraft:damage",
@@ -980,7 +980,7 @@ public class GenericTooltipTest {
                 false
         )).build(Lang.Branch.EXPLOSION), List.of(
                 "Explosion:",
-                "  -> Shape: STAR",
+                "  -> Shape: Star",
                 "  -> Colors: [1, 2, 3]",
                 "  -> Fade Colors: [3, 4, 5]",
                 "  -> Has Trail: true",
@@ -1013,8 +1013,8 @@ public class GenericTooltipTest {
                 "  -> Modifier:",
                 "    -> Id: minecraft:test",
                 "    -> Amount: 1.25",
-                "    -> Operation: ADD_VALUE",
-                "  -> Slot: HEAD"
+                "    -> Operation: Add Value",
+                "  -> Slot: Head"
         ));
     }
 
@@ -1028,7 +1028,7 @@ public class GenericTooltipTest {
                 "Attribute Modifier:",
                 "  -> Id: minecraft:test",
                 "  -> Amount: 1.25",
-                "  -> Operation: ADD_VALUE"
+                "  -> Operation: Add Value"
         ));
     }