+++ awi/common/src/test/java/com/yanny/awi/test/PlacementModifierTooltipTest.java
@@ -75,7 +75,7 @@ public class PlacementModifierTooltipTest {
                 32
         )).build(), List.of(
                 "Environment Scan Placement:",
-                "  -> Direction Of Search: UP",
+                "  -> Direction Of Search: Up",
                 "  -> Target Condition:",
                 "    -> Solid:",
                 "  -> Allowed Search Condition:",
@@ -88,7 +88,7 @@ public class PlacementModifierTooltipTest {
     public void testHeightmapPlacementTooltip() {
         assertTooltip(PlacementModifierTooltipUtils.getHeightmapPlacementTooltip(UTILS, HeightmapPlacement.onHeightmap(Heightmap.Types.MOTION_BLOCKING)).build(), List.of(
                 "Heightmap Placement:",
-                "  -> Heightmap: MOTION_BLOCKING"
+                "  -> Heightmap: Ground or Water Surface"
         ));
     }
 
@@ -159,17 +159,17 @@ public class PlacementModifierTooltipTest {
     public void testSurfaceRelativeThresholdFilterTooltip() {
         assertTooltip(PlacementModifierTooltipUtils.getSurfaceRelativeThresholdFilterTooltip(UTILS, SurfaceRelativeThresholdFilter.of(Heightmap.Types.WORLD_SURFACE_WG, Integer.MIN_VALUE, 5)).build(), List.of(
                 "Surface Relative Threshold Filter:",
-                "  -> Heightmap: WORLD_SURFACE_WG",
+                "  -> Heightmap: Highest Block, Plants Included",
                 "  -> Range: ≤5"
         ));
         assertTooltip(PlacementModifierTooltipUtils.getSurfaceRelativeThresholdFilterTooltip(UTILS, SurfaceRelativeThresholdFilter.of(Heightmap.Types.WORLD_SURFACE_WG, 3, Integer.MAX_VALUE)).build(), List.of(
                 "Surface Relative Threshold Filter:",
-                "  -> Heightmap: WORLD_SURFACE_WG",
+                "  -> Heightmap: Highest Block, Plants Included",
                 "  -> Range: ≥3"
         ));
         assertTooltip(PlacementModifierTooltipUtils.getSurfaceRelativeThresholdFilterTooltip(UTILS, SurfaceRelativeThresholdFilter.of(Heightmap.Types.WORLD_SURFACE_WG, 2, 8)).build(), List.of(
                 "Surface Relative Threshold Filter:",
-                "  -> Heightmap: WORLD_SURFACE_WG",
+                "  -> Heightmap: Highest Block, Plants Included",
                 "  -> Range: 2-8"
         ));
     }